imports/loading-import.html

<!DOCTYPE html>.
<html>
<head>  
<title>HTML Imports</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Hajime Morrita" href="mailto:morrita@google.com">
<link id="linkImport1" rel="import" href="resources/hello.html" />
<link id="linkImport2" rel="import" href="resources/hello.html" />
<link id="linkToTypicalHtml" rel="import" href="resources/typical.html" />
<link id="linkIncludingSubImport" rel="import" href="resources/sub-parent.html" />
</head>
<body>
<div id="log">Running test...</div>
<script>

test(function() {
  var linkImport1 = document.getElementById("linkImport1");
  var linkImport2 = document.getElementById("linkImport2");
  assert_equals(linkImport1.import, linkImport2.import);
  assert_true(linkImport1.import != null);
}, 
"Check on sharing imports",
{
  "help": ["https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#loading-imports"],
  "assert": "If LOCATION is already in the loaded import list, let FRAGMENT be the imported fragment for LOCATION and stop."
});

test(function() {
  var link = document.getElementById("linkToTypicalHtml");
  assert_equals(link.import.querySelector("p2").parentNode, 
                link.import.querySelector("parent"));
}, 
"Check on HTML Parsing",
{
  "help": ["https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#loading-imports"],
  "assert": "Let PARSER be a new HTML parser"
});

test(function() {
  var parent = document.getElementById("linkIncludingSubImport");
  var child = parent.import.querySelector("link").imports;
  assert_equals(child.title, "Hello");
}, 
"Check on sub-imports",
{
  "help": ["https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#loading-imports"],
  "assert": "Obtain the import resources using LINK and import fetching algorithm."
});

</script>
</body>
</html>
